requirements.txt
feedparser>=6.0
requests>=2.31
google-genai
tweepy>=4.14
python-dotenv>=1.0
python-dateutil>=2.8
beautifulsoup4>=4.12
google-generativeai>=0.8
